added recruitment and employee module for hr head

+++ WebApplication1/Models/Applicant.cs
@@ -9,17 +9,25 @@ namespace WebApplication1.Models
     using System.Data;
     using WebApplication1.Models.Supers;
 
+    public enum ApplicantStatus
+    {
+        Undecided = 1,
+        Accepted = 2,
+        Rejected = 3
+    }
+
     public class Applicant : ProfiledObject
     {
         public int ApplicantID { get; set; }
         public string Skills { get; set; }
         public string DesiredPosition { get; set; }
-       
+        public ApplicantStatus Status { get; set; }
+        
         public Applicant() {}
-        public Applicant(int ProfileID)
-            : base(ProfileID){}
+        public Applicant(int ProfileID, bool byPrimary = false)
+            : base(ProfileID, byPrimary){}
 
-        public override ProfiledObject FindProfile(int TableID, bool recursive = false, bool byPrimary = false)
+        public override ProfiledObject FindProfile(int TableID, bool recursive = true, bool byPrimary = false)
         {
             using (DataTable dt = this.DBHandler.Execute<DataTable>(CRUD.READ, "SELECT * FROM Applicant WHERE " + (byPrimary ? " ApplicantID " : " Profile ") + " = " + TableID))
             {
@@ -29,6 +37,7 @@ namespace WebApplication1.Models
                 this.ApplicantID = Int32.Parse(row["ApplicantID"].ToString());
                 this.Skills = row["Skills"].ToString();
                 this.DesiredPosition = row["DesiredPosition"].ToString();
+                this.Status = (ApplicantStatus)Int32.Parse(row["Status"].ToString());
 
                 if (recursive)
                 {
@@ -45,13 +54,14 @@ namespace WebApplication1.Models
             {
                 this.Profile.Create();
 
-                string columns = "INSERT INTO Applicant(Profile, Skills, DesiredPosition) OUTPUT INSERTED.ApplicantID";
-                string values = " Values(@Profile, @Skills, @DesiredPosition)";
+                string columns = "INSERT INTO Applicant(Profile, Skills, DesiredPosition, Status) OUTPUT INSERTED.ApplicantID";
+                string values = " Values(@Profile, @Skills, @DesiredPosition, @Status)";
                 Dictionary<string, dynamic> param = new Dictionary<string, dynamic>();
 
                 param.Add("@Profile", this.Profile.ProfileID);
                 param.Add("@Skills", this.Skills);
                 param.Add("@DesiredPositon", this.DesiredPosition);
+                param.Add("@Status", (int)this.Status);
 
                 this.ApplicantID = this.DBHandler.Execute<Int32>(CRUD.CREATE, columns + values, param);
             }
@@ -70,21 +80,24 @@ namespace WebApplication1.Models
                 this.Profile.Update(recursive);
             }
 
-            string set = "UPDATE Applicant SET Skills = @Skills AND "
-                         + "DesiredPosition = @DesiredPosition WHERE ApplicantID = " + this.ApplicantID;
+            string set = "UPDATE Applicant SET Skills = @Skills, "
+                         + "DesiredPosition = @DesiredPosition, Status = @Status OUTPUT INSERTED.ApplicantID WHERE ApplicantID = " + this.ApplicantID;
             Dictionary<string, dynamic> param = new Dictionary<string, dynamic>();
 
             param.Add("@Skills", this.Skills);
             param.Add("@DesiredPosition", this.DesiredPosition);
+            param.Add("@Status", (int)this.Status);
 
             this.DBHandler.Execute<Int32>(CRUD.UPDATE, set, param);
             return this;
         }
 
-        public Applicant Delete()
+        public Applicant Delete(bool recursive = false)
         {
             this.DBHandler.Execute<Int32>(CRUD.DELETE, "DELETE FROM Applicant WHERE ApplicantID = " + this.ApplicantID);
-            this.Profile.Delete();
+
+            if (recursive)
+                this.Profile.Delete();
             return this;
         }
     }
